requirements.txt
flask
joblib
scikit-learn
gunicorn
numpy
pandas
google-generativeai
python-dotenv